requirements_test.txt
pytest>=7.2.0
pytest-asyncio==0.24.0
pytest-homeassistant-custom-component==0.13.199
pytest-cov>=4.0.0
pytest-timeout>=2.1.0
pytest-xdist>=3.0.2
bleak>=0.20.0
bleak-retry-connector>=3.1.0